requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Web scraping
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3

# Data processing
pandas==2.2.3

# MCP Client (for GitHub integration)
mcp==1.3.0

# Async support
asyncio==3.4.3
aiohttp==3.11.11

# Utilities
python-dateutil==2.9.0

# Apify (for reliable job scraping)
apify-client==1.8.1

# Firecrawl (fastest MCP scraper, 7 sec avg, 83% accuracy)
firecrawl-py==4.6.0

# PDF and AI (REQUIRED for job matching and resume parsing)
PyPDF2==3.0.1
google-generativeai==0.8.3
